oct-smart-tutor/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.18.0
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.9.2
